+++ clang-tools-extra/clang-tidy/misrac/Rule8p12Check.h
@@ -1,0 +1,40 @@
+//===--- Rule8p12Check.h - clang-tidy ---------------------------*- C++ -*-===//
+//
+// Part of the LLVM Project, under the Apache License v2.0 with LLVM Exceptions.
+// See https://llvm.org/LICENSE.txt for license information.
+// SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
+//
+//===----------------------------------------------------------------------===//
+
+#ifndef LLVM_CLANG_TOOLS_EXTRA_CLANG_TIDY_MISRAC_RULE8P12CHECK_H
+#define LLVM_CLANG_TOOLS_EXTRA_CLANG_TIDY_MISRAC_RULE8P12CHECK_H
+
+#include "MisracBaseCheck.h"
+
+namespace clang::tidy::misrac {
+
+/// Detects issues in accordance with MISRA C rule 8p12:
+/// "Within an enumerator list, the value of an implicitly-specified
+/// enumeration constant shall be unique".
+///
+///
+/// For the user-facing documentation see:
+/// http://clang.llvm.org/extra/clang-tidy/checks/misrac/rule-8p12.html
+class Rule8p12Check : public MisracBaseCheck {
+public:
+  Rule8p12Check(StringRef Name, ClangTidyContext *Context)
+      : MisracBaseCheck(Name, Context) {}
+
+protected:
+  void addMatchers(ast_matchers::MatchFinder *Finder) override;
+  void dispatch(const ast_matchers::MatchFinder::MatchResult &Result) override;
+
+private:
+  void dispatch(const EnumDecl *Decl);
+
+  llvm::DenseMap<llvm::APInt, const EnumConstantDecl *> Map;
+};
+
+} // namespace clang::tidy::misrac
+
+#endif // LLVM_CLANG_TOOLS_EXTRA_CLANG_TIDY_MISRAC_RULE8P12CHECK_H